dtype

+++ vla/model/framework/VLM4A/QwenGR00T.py
@@ -316,7 +316,7 @@ if __name__ == "__main__":
 
     batch = [sample, sample2]
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    model = model.to(device)
+    model = model.to(device,dtype=torch.bfloat16)
     forward_output = model(batch)
     action_loss = forward_output["action_loss"]
     print(f"\33[93mAction Loss: {action_loss.item()}\33[0m")
